apps/backend-rag/requirements.txt
# ZANTARA RAG Backend Dependencies v5.3 (Ultra Hybrid) - FIXED
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Vector Database
qdrant-client==1.7.0

# Google Cloud Integration (Core for v5.3 Hybrid)
google-generativeai==0.3.2
google-api-python-client==2.100.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-cloud-storage==2.10.0

# AI/ML
openai==1.3.0
anthropic==0.7.8
openpyxl==3.1.2
email-validator==2.1.0
pypdf==3.17.1
EbookLib==0.17.1
beautifulsoup4==4.12.2

# LangChain Ecosystem (Minimal - only used dependencies)
langgraph==0.0.26
langchain-core>=0.1.52,<0.2.0
# Note: langchain, langsmith, langchain-text-splitters removed - not used in codebase

# HTTP/Network
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1  # Required for personality_service.py

# Authentication
python-jose[cryptography]==3.3.0
bcrypt==4.0.1

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9  # Required for agents/conversation_trainer.py
redis==5.0.1
sqlmodel==0.0.14  # SQLModel for type-safe database models (Prime Standard)

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
pytz==2023.3

# Development & Quality Tools
pre-commit==3.6.0  # Git hooks automation
ruff==0.1.8  # Fast Python linter and formatter
